AWS/requirements.txt
aniso8601==8.0.0
atomicwrites==1.4.0
attrs==21.4.0
bip-utils==1.0.5
bleach==5.0.0
boto3==1.23.0
botocore==1.26.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
coverage==6.3.3
cryptography==37.0.2
cytoolz==0.11.2
docutils==0.18.1
ecdsa==0.16.1
eth-abi==2.1.1
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-typing==2.3.0
eth-utils==1.10.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.0
keyring==23.5.0
MarkupSafe==1.1.1
mnemonic==0.19
packaging==21.3
parsimonious==0.8.1
pkginfo==1.8.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pysha3==1.0.2
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
rlp==1.2.0
s3transfer==0.5.2
six==1.15.0
SQLAlchemy==1.3.18
thor-devkit==1.0.12
thor-requests==1.2.2
tomli==2.0.1
toolz==0.11.2
twine==4.0.0
typing_extensions==4.2.0
urllib3==1.26.9
voluptuous==0.12.0
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.8.0